make PR improvements
