+++ battles/helpers/fight.py
@@ -1,8 +1,6 @@
 # rounds figth
-def compare_hitpoints(pokemon_1, pokemon_2):
-    if pokemon_1.hitpoints > pokemon_2.hitpoints:
-        return pokemon_1
-    return pokemon_2
+def get_pokemon_with_highest_hitpoints(pokemon_1, pokemon_2):
+    return pokemon_1 if pokemon_1.hitpoints > pokemon_2.hitpoints else pokemon_2
 
 
 def compare_attack_defense(pokemon_1, pokemon_2):
@@ -14,7 +12,7 @@ def compare_attack_defense(pokemon_1, pokemon_2):
 def get_pokemon_round_winner(pokemon_1, pokemon_2):
     if compare_attack_defense(pokemon_1, pokemon_2) == compare_attack_defense(pokemon_2, pokemon_1):
         return compare_attack_defense(pokemon_1, pokemon_2)
-    return compare_hitpoints(pokemon_1, pokemon_2)
+    return get_pokemon_with_highest_hitpoints(pokemon_1, pokemon_2)
 
 
 # battle fight
